Prioridad 1,2 o 3, validaciones, diseño 1.3 form

+++ GrupoF.TP.CAI.Pampazon/Formularios/1. Registrar Orden de Preparación/Modelos/BuscarClienteModel.cs
@@ -177,7 +177,7 @@ namespace GrupoF.TP.CAI.Pampazon.Modelos
                     CodigoCliente = orden.CodigoCliente,
                     CodigoTransportista = orden.CodigoTransportista,
                     EstadoOrden = Estados.Estado.Pendiente,
-                    Prioridad = (Prioridad)orden.Prioridad,
+                    Prioridad = (Prioridad)(int)orden.Prioridad,
 
                     Detalle = orden.ProductosOrden.Select(p => new OrdenDePreparacionDetalle
                     {
